move `get_minimum_supported_version` to where it's used
